Move the phoenix files from hexdocs
I had tried learning through hexdocs because the book was out of date
(or so I thought). I ran into the same error reading their guide and
then figured it out. The book should be good enough. The phoenix website
gets very vague very quickly. It seems as though it was written by the
phoenix creators, who already understand plug and other modules. Not
having that background makes it very hard to read those docs. I need a
book which focuses on actually teaching the concepts.